Add some debug log and fix the recordCnt (#10)
---

*Motivation*

The recordCnt is used to control commit action. When we
write a record successfully, the recordCnt should be increased.

The sink configuration for hudi can not return an default class,
that will cause the config validation failed. We need to use
jsonMapper to parse it then the type will be load.

*Modifications*

- add more debug info
- fix some issues when running the hudi sink